pyproject.toml
[project]
name = "ngo-app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "bcrypt==4.2.0",
    "geopy==2.4.1",
    "google-generativeai>=0.3.0",
    "pandas==2.2.3",
    "psycopg2-binary>=2.9.9",
    "python-dateutil==2.9.0.post0",
    "python-dotenv>=1.0.0",
    "sqlalchemy==2.0.36",
    "streamlit==1.39.0",
]
